Add authorization tests. Fix several bugs.
`ouser_permission_level` null literal:
If a user's ouser_permission_level field is null literal and the
permissions for an object or method are an integer, that user was
granted access. This fixes that bug.

Basic auth bug:
If a request had no authenticated session but provided basic auth
credentials and the permissions for an object or method are an integer,
that user was granted access. This fixes that bug.

Graduated permissions wrong logic:
If the constant `__OBRAY_GRADUATED_PERMISSIONS__` is defined and the
permissions for an object or method are an integer, the framework
checks if the user's `ouser_permission_level` is GREATER THAN the
required level, and if so DENIES the user access. This corrects the
behavior to check i the user's `ouser_permission_level` is LESS THAN
the required level, and if so DENIES the user access.